requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.7
numpy>=1.26.0
pandas>=2.1.0
tabulate>=0.9.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
wandb>=0.17.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
streamlit>=1.32.0
requests>=2.31.0
